requirements.txt
faster-whisper>=1.0.0
huggingface-hub>=0.20.0
transformers>=4.36.0
torch>=2.1.0
numpy>=1.24.0
ffmpeg-python>=0.2.0
protobuf>=4.25.1
packaging